backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
slowapi==0.1.9
pydantic==2.7.0
pydantic-settings==2.3.0
email-validator==2.1.1

# ML
torch==2.5.1
torchvision==0.20.1
timm==1.0.12
facenet-pytorch==2.5.3
onnxruntime==1.19.0
albumentations==1.4.20
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.9.0
seaborn==0.13.2
tensorboard==2.17.0